+++ src/app/containers/index.js
@@ -3,3 +3,4 @@ export { default as HeaderContainer } from './HeaderContainer';
 export { default as FieldContainer } from './FieldContainer';
 export { default as MetamaskButtonContainer } from './metamask/MetamaskButtonContainer';
 export { default as MetamaskFormContainer } from './metamask/MetamaskFormContainer';
+export { default as LanguageSelectContainer } from './LanguageSelectContainer';
