+++ src/app/tabs/user/components/index.js
@@ -1,2 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
 export { default as UserTabComponent } from './UserTabComponent';
-export { default as LanguageSelectComponent } from './LanguageSelectComponent';
