+++ src/app/tabs/user/containers/index.js
@@ -1,2 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
 export { default as UserTabContainer } from './UserTabContainer';
-export { default as LanguageSelectContainer } from './LanguageSelectContainer';
